Implementing User Authorization

+++ server/routes/posts.js
@@ -38,6 +38,7 @@ router.post(
       title: req.body.title,
       content: req.body.content,
       imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId,
     });
     post.save().then((createdPost) => {
       res.status(201).json({
@@ -92,10 +93,20 @@ router.put(
       title: req.body.title,
       content: req.body.content,
       imagePath: imagePath,
+      creator: req.userData.userId
     });
     console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then((result) => {
-      res.status(200).json({ message: "Post has been updated successfully" });
+    Post.updateOne(
+      { _id: req.params.id, creator: req.userData.userId },
+      post
+    ).then((result) => {
+      console.log(result);
+      if(result.nModified > 0) {
+        res.status(200).json({ message: "Post has been updated successfully" });
+      } else {
+        res.status(401).json({ message: "Not Authorized" });
+      }
+      
     });
   }
 );
@@ -111,8 +122,13 @@ router.get("/:id", (req, res, err) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, err) => {
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
+    if(result.n > 0) {
+      res.status(200).json({ message: "Post deleted!" });
+    } else {
+      res.status(401).json({ message: "Not Authorized" });
+    }
+    
   });
 });
